[Bronze V] Title: 이상한 기호, Time: 172 ms, Memory: 17656 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,8 +5,8 @@ public class Main {
 
         Scanner sc = new Scanner(System.in);
 
-        int A = sc.nextInt();
-        int B = sc.nextInt();
+        long A = sc.nextLong();
+        long B = sc.nextLong();
 
         System.out.println((A + B) * (A - B));
